Refactors AccordionForm into an independent component (#60334)
* Refactors AccordionForm into an independent component

* Rename useSections to sectionGenerator
