+++ client/signup/steps/site-info-collection/index.tsx
@@ -1,6 +1,16 @@
 import { useTranslate } from 'i18n-calypso';
-import { ChangeEvent, useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
+import AccordionForm from 'calypso/signup/accordion-form/accordion-form';
+import {
+	ContactInformation,
+	TextInputField,
+	SocialMediaProfiles,
+} from 'calypso/signup/accordion-form/form-components';
+import {
+	AccordionSectionProps,
+	sectionGeneratorReturnType,
+} from 'calypso/signup/accordion-form/types';
 import StepWrapper from 'calypso/signup/step-wrapper';
 import { saveSignupStep } from 'calypso/state/signup/progress/actions';
 import {
@@ -12,9 +22,6 @@ import {
 	getSiteInfoCollectionData,
 	getSiteInfoCollectionCurrentIndex,
 } from 'calypso/state/signup/steps/site-info-collection/selectors';
-import { AccordionForm } from './accordion-form';
-import { ContactInformation, TextInputField, SocialMediaProfiles } from './form-components';
-import { AccordionSectionProps, ValidationErrors, ValidatorFunction } from './types';
 
 import './style.scss';
 
@@ -25,6 +32,117 @@ interface SiteInformationCollectionProps {
 	goToNextStep: () => void;
 }
 
+function sectionGenerator() {
+	return ( {
+		translate,
+		formValues,
+		formErrors,
+		onChangeField,
+	}: sectionGeneratorReturnType< SiteInfoCollectionData > ) => {
+		const sections: AccordionSectionProps< SiteInfoCollectionData >[] = [
+			{
+				title: translate( '%d. Name of your business', {
+					args: [ 1 ],
+					comment: 'This is the serial number: 1',
+				} ),
+				component: (
+					<TextInputField
+						label={ translate(
+							'Please enter the name of your business as you want it to appear on your new site.'
+						) }
+						value={ formValues.siteTitle }
+						error={ formErrors.siteTitle }
+						name="siteTitle"
+						onChange={ onChangeField }
+					/>
+				),
+				showSkip: false,
+				summary: formValues.siteTitle,
+				validate: ( formValues ) => {
+					const isValid = Boolean( formValues.siteTitle?.length );
+					return {
+						result: isValid,
+						errors: {
+							siteTitle: isValid ? null : translate( 'Please enter a valid site title.' ),
+						},
+					};
+				},
+			},
+			{
+				title: translate( '%d. Site Description', {
+					args: [ 2 ],
+					comment: 'This is the serial number: 2',
+				} ),
+				component: (
+					<TextInputField
+						label={ translate( 'A short description for your website.' ) }
+						value={ formValues.siteDescription }
+						name="siteDescription"
+						onChange={ onChangeField }
+					/>
+				),
+				showSkip: true,
+				summary: formValues.siteDescription,
+			},
+			{
+				title: translate( '%d. Social Media Profiles', {
+					args: [ 3 ],
+					comment: 'This is the serial number: 3',
+				} ),
+				component: (
+					<SocialMediaProfiles
+						facebookProps={ {
+							value: formValues.facebookUrl,
+							name: 'facebookUrl',
+						} }
+						twitterProps={ {
+							value: formValues.twitterUrl,
+							name: 'twitterUrl',
+						} }
+						linkedinProps={ {
+							value: formValues.linkedinUrl,
+							name: 'linkedinUrl',
+						} }
+						instagramProps={ {
+							value: formValues.instagramUrl,
+							name: 'instagramUrl',
+						} }
+						onChange={ onChangeField }
+					/>
+				),
+				showSkip: true,
+				summary: undefined,
+			},
+			{
+				title: translate( '%d. Contact Information', {
+					args: [ 4 ],
+					comment: 'This is the serial number: 4',
+				} ),
+				component: (
+					<ContactInformation
+						displayEmailProps={ {
+							value: formValues.displayEmail,
+							name: 'displayEmail',
+						} }
+						displayPhoneProps={ {
+							value: formValues.displayPhone,
+							name: 'displayPhone',
+						} }
+						displayAddressProps={ {
+							value: formValues.displayAddress,
+							name: 'displayAddress',
+						} }
+						onChange={ onChangeField }
+					/>
+				),
+				showSkip: true,
+				summary: undefined,
+			},
+		];
+		return sections;
+	};
+}
+
 function SiteInformationCollection( {
 	additionalStepData,
 	stepName,
@@ -32,7 +150,6 @@ function SiteInformationCollection( {
 	goToNextStep,
 }: SiteInformationCollectionProps ) {
 	const dispatch = useDispatch();
-	const translate = useTranslate();
 
 	useEffect( () => {
 		dispatch( saveSignupStep( { stepName } ) );
@@ -43,151 +160,7 @@ function SiteInformationCollection( {
 	);
 	const currentIndex: number = useSelector( getSiteInfoCollectionCurrentIndex );
 
-	// Initialize local state with the values from the redux store
-	const [ formValues, setFormValues ] = useState< SiteInfoCollectionData >(
-		siteInfoCollectionDataFromStore
-	);
-	const onChangeField = ( { target: { name, value } }: ChangeEvent< HTMLInputElement > ) => {
-		setFormValues( {
-			...formValues,
-			[ name ]: value,
-		} );
-	};
-
-	const [ formErrors, setFormErrors ] = useState< ValidationErrors >( {} );
-
-	const sections: AccordionSectionProps[] = [
-		{
-			title: translate( '%d. Name of your business', {
-				args: [ 1 ],
-				comment: 'This is the serial number: 1',
-			} ),
-			component: (
-				<TextInputField
-					label={ translate(
-						'Please enter the name of your business as you want it to appear on your new site.'
-					) }
-					value={ formValues.siteTitle }
-					error={ formErrors.siteTitle }
-					name="siteTitle"
-					onChange={ onChangeField }
-				/>
-			),
-			showSkip: false,
-			summary: formValues.siteTitle,
-			validate: ( { siteTitle } ) => {
-				const isValid = Boolean( siteTitle?.length );
-				return {
-					result: isValid,
-					errors: {
-						siteTitle: isValid ? null : translate( 'Please enter a valid site title.' ),
-					},
-				};
-			},
-		},
-		{
-			title: translate( '%d. Site Description', {
-				args: [ 2 ],
-				comment: 'This is the serial number: 2',
-			} ),
-			component: (
-				<TextInputField
-					label={ translate( 'A short description for your website.' ) }
-					value={ formValues.siteDescription }
-					name="siteDescription"
-					onChange={ onChangeField }
-				/>
-			),
-			showSkip: true,
-			summary: formValues.siteDescription,
-		},
-		{
-			title: translate( '%d. Social Media Profiles', {
-				args: [ 3 ],
-				comment: 'This is the serial number: 3',
-			} ),
-			component: (
-				<SocialMediaProfiles
-					facebookProps={ {
-						value: formValues.facebookUrl,
-						name: 'facebookUrl',
-					} }
-					twitterProps={ {
-						value: formValues.twitterUrl,
-						name: 'twitterUrl',
-					} }
-					linkedinProps={ {
-						value: formValues.linkedinUrl,
-						name: 'linkedinUrl',
-					} }
-					instagramProps={ {
-						value: formValues.instagramUrl,
-						name: 'instagramUrl',
-					} }
-					onChange={ onChangeField }
-				/>
-			),
-			showSkip: true,
-			summary: undefined,
-		},
-		{
-			title: translate( '%d. Contact Information', {
-				args: [ 4 ],
-				comment: 'This is the serial number: 4',
-			} ),
-			component: (
-				<ContactInformation
-					displayEmailProps={ {
-						value: formValues.displayEmail,
-						name: 'displayEmail',
-					} }
-					displayPhoneProps={ {
-						value: formValues.displayPhone,
-						name: 'displayPhone',
-					} }
-					displayAddressProps={ {
-						value: formValues.displayAddress,
-						name: 'displayAddress',
-					} }
-					onChange={ onChangeField }
-				/>
-			),
-			showSkip: true,
-			summary: undefined,
-		},
-	];
-
-	const isSectionAtIndexTouchedInitialState: Record< string, boolean > = {};
-	for ( let i = 0; i <= currentIndex; i++ ) {
-		isSectionAtIndexTouchedInitialState[ `${ i }` ] = true;
-	}
-
-	const [ isSectionAtIndexTouched, setIsSectionAtIndexTouched ] = useState<
-		Record< string, boolean >
-	>( isSectionAtIndexTouchedInitialState );
-
-	const runValidatorAndSetFormErrors = ( validator: ValidatorFunction ) => {
-		const validationResult = validator( formValues );
-		setFormErrors( {
-			...formErrors,
-			...validationResult.errors,
-		} );
-		return validationResult;
-	};
-
-	const submitStep = () => {
-		// Re-run validation on all sections before submitting
-		for ( let index = 0; index < sections.length; index++ ) {
-			const section = sections[ index ];
-			if ( section.validate ) {
-				const validationResult = runValidatorAndSetFormErrors( section.validate );
-				if ( ! validationResult.result ) {
-					dispatch( updateSiteInfoCurrentIndex( index ) );
-					return;
-				}
-			}
-		}
-
+	const onSubmit = ( formValues: SiteInfoCollectionData ) => {
 		const step = {
 			stepName,
 			...additionalStepData,
@@ -198,40 +171,18 @@ function SiteInformationCollection( {
 		goToNextStep();
 	};
 
-	const onOpen = ( currentIndex: number ) => {
-		dispatch( updateSiteInfoCurrentIndex( currentIndex ) );
-		dispatch( updateSiteInfoValues( formValues ) );
-		setIsSectionAtIndexTouched( { ...isSectionAtIndexTouched, [ `${ currentIndex }` ]: true } );
-	};
-
-	const onNext = ( validator?: ValidatorFunction ) => {
-		dispatch( updateSiteInfoValues( formValues ) );
-
-		if ( validator ) {
-			const validationResult = runValidatorAndSetFormErrors( validator );
-			if ( ! validationResult.result ) {
-				return;
-			}
-		}
-
-		if ( currentIndex < sections.length - 1 ) {
-			dispatch( updateSiteInfoCurrentIndex( currentIndex + 1 ) );
-			setIsSectionAtIndexTouched( {
-				...isSectionAtIndexTouched,
-				[ `${ currentIndex + 1 }` ]: true,
-			} );
-		} else {
-			submitStep();
-		}
-	};
-
 	return (
 		<AccordionForm
-			sections={ sections }
+			sectionGenerator={ sectionGenerator }
+			formValuesInitialState={ siteInfoCollectionDataFromStore }
 			currentIndex={ currentIndex }
-			isSectionAtIndexTouched={ isSectionAtIndexTouched }
-			onNext={ onNext }
-			onOpen={ onOpen }
+			updateCurrentIndex={ ( index ) => {
+				dispatch( updateSiteInfoCurrentIndex( index ) );
+			} }
+			updateFormValues={ ( formValues ) => {
+				dispatch( updateSiteInfoValues( formValues ) );
+			} }
+			onSubmit={ onSubmit }
 		/>
 	);
 }
